moved IFileSysTreeItem to shared lib and referenced in vs-code-designer
