+++ apps/Standalone/src/dataMapperV1/app/DataMapperStandaloneDesignerV2.tsx
@@ -4,7 +4,7 @@ import {
   DataMapperDesigner as DataMapperDesignerV2,
   DataMapDataProvider as DataMapDataProviderV2,
   DataMapperDesignerProvider as DataMapperDesignerProviderV2,
-  IDataMapperFileService,
+  type IDataMapperFileService,
 } from '@microsoft/logic-apps-data-mapper-v2';
 import type { AppDispatch, RootState } from '../state/Store';
 import { AzureThemeDark } from '@fluentui/azure-themes/lib/azure/AzureThemeDark';
@@ -16,7 +16,7 @@ import { InitDataMapperApiService, defaultDataMapperApiServiceOptions, getFuncti
 import { Theme as ThemeType } from '@microsoft/logic-apps-shared';
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import type { IFileSysTreeItem } from '@microsoft/logic-apps-data-mapper-v2/src/models/Tree';
+import type { IFileSysTreeItem } from '@microsoft/logic-apps-shared';
 
 const mockFileItems: IFileSysTreeItem[] = [
   {
@@ -46,16 +46,13 @@ class DataMapperFileService implements IDataMapperFileService {
   private verbose: boolean;
 
   constructor(verbose: boolean) {
-    this.verbose = verbose
+    this.verbose = verbose;
   }
 
-  public saveMapDefinitionCall = (
-    dataMapDefinition: string,
-    mapMetadata: string
-  ) => {
+  public saveMapDefinitionCall = (dataMapDefinition: string, mapMetadata: string) => {
     if (this.verbose) {
-      console.log('Saved definition: ' + dataMapDefinition)
-      console.log('Saved metadata: ' + mapMetadata)
+      console.log(`Saved definition: ${dataMapDefinition}`);
+      console.log(`Saved metadata: ${mapMetadata}`);
     }
   };
 
@@ -63,8 +60,6 @@ class DataMapperFileService implements IDataMapperFileService {
     return;
   };
 }
-
-
 
 const customXsltPath = ['folder/file.xslt', 'file2.xslt'];
