+++ apps/vs-code-react/src/run-service/types.ts
@@ -1,6 +1,5 @@
-import type { IFileSysTreeItem } from '@microsoft/logic-apps-data-mapper-v2/src/models/Tree';
 import type { InitializePayload, Status } from '../state/WorkflowSlice';
-import type { ApiHubServiceDetails, SchemaType } from '@microsoft/logic-apps-shared';
+import type { ApiHubServiceDetails, SchemaType, IFileSysTreeItem } from '@microsoft/logic-apps-shared';
 import type { MapDefinitionData, ExtensionCommand, ConnectionsData, IDesignerPanelMetadata } from '@microsoft/vscode-extension-logic-apps';
 
 export interface IApiService {
