--- libs/data-mapper-v2/src/models/Tree.ts
@@ -1,13 +1,0 @@
-export type IFileSysTreeItem = ITreeDirectory | ITreeFile;
-
-export interface ITreeDirectory {
-  name: string;
-  type: 'directory';
-  children: IFileSysTreeItem[];
-}
-
-export interface ITreeFile {
-  name: string;
-  type: 'file';
-  fullPath?: string;
-}
