+++ libs/vscode-extension/src/lib/models/dataMapper.ts
@@ -1,5 +1,5 @@
 import type { ExtensionCommand } from './extensioncommand';
-import type { MapDefinitionEntry, MapMetadata, SchemaType } from '@microsoft/logic-apps-shared';
+import type { IFileSysTreeItem, MapDefinitionEntry, MapMetadata, SchemaType } from '@microsoft/logic-apps-shared';
 
 type InitializeData = { project: string };
 type FetchSchemaData = { fileName: string; type: SchemaType };
@@ -18,7 +18,7 @@ export type MessageToWebview =
   | { command: typeof ExtensionCommand.fetchSchema; data: FetchSchemaData }
   | { command: typeof ExtensionCommand.loadDataMap; data: MapDefinitionData }
   | { command: typeof ExtensionCommand.showAvailableSchemas; data: string[] }
-  | { command: typeof ExtensionCommand.showAvailableSchemasV2; data: any[] } //todo: change to IFileSysTreeItem
+  | { command: typeof ExtensionCommand.showAvailableSchemasV2; data: IFileSysTreeItem[] }
   | { command: typeof ExtensionCommand.getAvailableCustomXsltPaths; data: string[] }
   | { command: typeof ExtensionCommand.setXsltData; data: XsltData }
   | { command: typeof ExtensionCommand.setRuntimePort; data: string }
